backend/requirements.txt
Django==4.2.20
Pillow==11.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
